REQUISITES: Allow reserve/cancel requests concurrently with GET report request.
To do it, we need to avoid giant mutex lock.
Under the situation, to avoid race conditions, we take counts of
reservations before request and after response of GET report API.

Please note that reserve/cancel requests would be issued right after the
respose of GET report API and right before taking the count of reservations.